Fix image uploading issues and optimizing images

+++ backend/src/users/users.controller.ts
@@ -12,6 +12,7 @@ import { UsersGateway } from './users.gateway';
 import { get } from 'http';
 import { Response } from 'express';
 import * as fs from 'fs';
+import { PrismaService } from 'src/prisma/prisma.service';
 // import { clearConfigCache } from 'prettier';
 
 @Controller()
@@ -20,6 +21,7 @@ export class UsersController {
         private jwt: JwtService,
         private userservice: UsersService,
         private usergtw: UsersGateway,
+        private prisma: PrismaService,
     ) { }
 
     @UseGuards(JwtAuthGuard)
@@ -112,6 +114,10 @@ export class UsersController {
                 if (extname(avatar.originalname) !== '.png' && extname(avatar.originalname) !== '.jpg' && extname(avatar.originalname) !== '.jpeg' && extname(avatar.originalname) !== '.gif') {
                     return cb(new HttpException('Only images are allowed', HttpStatus.BAD_REQUEST), '')
                 }
+                // check size of file
+                if (avatar.size > 1024 * 1024) {
+                    return cb(new HttpException('File too large', HttpStatus.BAD_REQUEST), '');
+                }
                 if (fs.existsSync(`./uploads/${Name}.jpeg`)) {
                     fs.unlinkSync(`./uploads/${Name}.jpeg`);
                 }
@@ -121,6 +127,19 @@ export class UsersController {
     }))
     async uploadFile(@UploadedFile(
     ) file: Express.Multer.File, @Req() req) {
+        if (file) {
+            let url = `http://localhost:4000/${req.user.id}.jpeg`;
+            await this.prisma.user.update({
+                where: {
+                    id: req.user.id,
+                },
+                data: {
+                    picture: url,
+                },
+            });
+        } else {
+            throw new HttpException('Avatar is required', HttpStatus.BAD_REQUEST);
+        }
         return file;
     }
